+++ src/main/kotlin/com/algocompiler/Token.kt
@@ -32,6 +32,8 @@ enum class TokenType {
     PROCEDURE,
     RETOURNER,
     TABLEAU,
+    VAR,
+    PAS,
 
     // Types
     ENTIER,